requirements.txt
flask
langchain
langchain-google-genai
langchain-google-vertexai
langchain-community
langchain-openai
beautifulsoup4
google-cloud-storage
python-dotenv
google-cloud-aiplatform
openai
google-generativeai
aiohttp
langgraph
tiktoken
gunicorn==22.0.0
Werkzeug==3.0.3
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
